requirements.txt
pytest~=7.4.3
pytest-cov
logassert
Appium-Python-Client~=3.1.0